requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
PyYAML==6.0.1
rich==13.7.0
jinja2==3.1.2
pytest==7.4.3
click==8.1.7
aiofiles==23.2.1
